M🅱️ve meme c🅱️mmands t🅱️ their 🅱️wn m🅱️dule
🅱️h, and I als🅱️ fixed bm🅱️ji
