+++ tg_bot/modules/special.py
@@ -8,7 +8,7 @@ from telegram.ext import MessageHandler, Filters, CommandHandler
 from telegram.ext.dispatcher import run_async
 from tg_bot.modules.helper_funcs.chat_status import is_user_ban_protected
 
-import telegram, random
+import telegram
 import tg_bot.modules.sql.users_sql as sql
 from tg_bot import dispatcher, OWNER_ID, LOGGER
 from tg_bot.modules.helper_funcs.filters import CustomFilters
@@ -88,51 +88,6 @@ def chats(bot: Bot, update: Update):
         update.effective_message.reply_document(document=output, filename="chatlist.txt",
                                                 caption="Here is the list of chats in my database.")
 
-# D A N K module by @deletescape - based on https://github.com/wrxck/mattata/blob/master/plugins/copypasta.mattata
-
-@run_async
-def copypasta(bot: Bot, update: Update):
-    message = update.effective_message
-    emojis = ["😂", "😂", "👌", "✌", "💞", "👍", "👌", "💯", "🎶", "👀", "😂", "👓", "👏", "👐", "🍕", "💥", "🍴", "💦", "💦", "🍑", "🍆", "😩", "😏", "👉👌", "👀", "👅", "😩", "🚰"]
-    reply_text = random.choice(emojis)
-    b_char = random.choice(message.reply_to_message.text).lower() # choose a random character in the message to be substituted with 🅱️
-    for c in message.reply_to_message.text:
-        if c == " ":
-            reply_text += random.choice(emojis)
-        elif c in emojis:
-            reply_text += c
-            reply_text += random.choice(emojis)
-        elif c.lower() == b_char:
-            reply_text += "🅱️"
-        else:
-            if bool(random.getrandbits(1)):
-                reply_text += c.upper()
-            else:
-                reply_text += c.lower()
-    reply_text += random.choice(emojis)
-    message.reply_to_message.reply_text(reply_text)
-	
-# D A N K module by @deletescape 
-
-@run_async
-def bmoji(bot: Bot, update: Update):
-    message = update.effective_message
-    b_char = random.choice(message.reply_to_message.text).lower() # choose a random character in the message to be substituted with 🅱️
-    reply_text = message.replace(b_char, "🅱️").replace(b_char.upper(), "🅱️")
-    message.reply_to_message.reply_text(reply_text)
-
-@run_async
-def clapmoji(bot: Bot, update: Update):
-    message = update.effective_message
-    reply_text = "👏 "
-    for i in message.reply_to_message.text:
-        if i == " ":
-            reply_text += " 👏 "
-        else:
-            reply_text += i
-    reply_text += " 👏"
-    message.reply_to_message.reply_text(reply_text)
-
 __help__ = ""  # no help string
 
 __mod_name__ = "Special"
@@ -142,19 +97,9 @@ SNIPE_HANDLER = CommandHandler("snipe", snipe, pass_args = True, filters=CustomF
 BANALL_HANDLER = CommandHandler("banall", banall, pass_args = True, filters=Filters.user(OWNER_ID))
 QUICKSCOPE_HANDLER = CommandHandler("quickscope", quickscope, pass_args = True, filters=CustomFilters.sudo_filter)
 QUICKUNBAN_HANDLER = CommandHandler("quickunban", quickunban, pass_args = True, filters=CustomFilters.sudo_filter)
-COPYPASTA_HANDLER = DisableAbleCommandHandler("copypasta", copypasta)
-COPYPASTA_ALIAS_HANDLER = DisableAbleCommandHandler("😂", copypasta)
-CLAPMOJI_HANDLER = DisableAbleCommandHandler("clapmoji", clapmoji)
-CLAPMOJI_ALIAS_HANDLER = DisableAbleCommandHandler("👏", clapmoji)
-BMOJI_HANDLER = DisableAbleCommandHandler("🅱️", bmoji)
 
 dispatcher.add_handler(CHATSS_HANDLER)
 dispatcher.add_handler(SNIPE_HANDLER)
 dispatcher.add_handler(BANALL_HANDLER)
 dispatcher.add_handler(QUICKSCOPE_HANDLER)
 dispatcher.add_handler(QUICKUNBAN_HANDLER)
-dispatcher.add_handler(COPYPASTA_HANDLER)
-dispatcher.add_handler(COPYPASTA_ALIAS_HANDLER)
-dispatcher.add_handler(CLAPMOJI_HANDLER)
-dispatcher.add_handler(CLAPMOJI_ALIAS_HANDLER)
-dispatcher.add_handler(BMOJI_HANDLER)
